Stage 1
Turns <App /> into a stateful component to allow for controls in the <Counter /> to increment and decrement the values. Tests also added for 100% coverage.

Note: This would not have worked correctly unless the data was managed by state somewhere, the functions would have modified the data but the values would not have been changed on screen because a render would not have been triggered. 


Fixed tests